BROKEN: add check that panics appropriately when setting storage
Mostly this already does the right thing: it encloses the `host_storage_set`
function with a context providing the current block. That way, if it
looks like someone is trying to update the code section, we can panic
**unless** polkadot has already included a magic key, `UPGRADE_BLOCK_KEY`,
in the storage context which has a lower block number than the current block.

Unfortunately, `replace_implementation` is not generic over Fn;
it instead expects a literal function pointer. If this looks like a
good approach, the next step is going to be to update that function
such that it is in fact generic, so it'll accept the generated
closure here.

One of the reasons I don't yet know whether or not this is a plausible
approach is that I don't know what WASM's capabilities are; generating
temporary closures and passing them around is a non-trivial capability,
and still trips people up in i.e. golang.