BROKEN: add current block number to parachain validation call
The first and obvious response to this implemention is "Why are you
changing the function signature? Just add it to the ValidationParams!"

The problem with that approach is that leads to a circular dependency:
to add the block number to the ValidationParams struct, we need to
add a dependency from the ValidationParams definition

https://github.com/paritytech/polkadot/blob/5ed74bac90ac24a2abd9828e310caaccde26f15c/parachain/src/lib.rs#L61-L68

to the block number definiton:

https://github.com/paritytech/polkadot/blob/5ed74bac90ac24a2abd9828e310caaccde26f15c/primitives/src/lib.rs#L30-L32

Adding that dependency causes this error:

```
error: cyclic package dependency: package `polkadot-parachain v0.7.21 (/home/coriolinus/Documents/Projects/paritytech/polkadot/parachain)` depends on itself. Cycle:
package `polkadot-parachain v0.7.21 (/home/coriolinus/Documents/Projects/paritytech/polkadot/parachain)`
    ... which is depended on by `polkadot-primitives v0.7.21 (/home/coriolinus/Documents/Projects/paritytech/polkadot/primitives)`
    ... which is depended on by `polkadot-parachain v0.7.21 (/home/coriolinus/Documents/Projects/paritytech/polkadot/parachain)`
    ... which is depended on by `adder v0.7.21 (/home/coriolinus/Documents/Projects/paritytech/polkadot/test-parachains/adder)`
```

For now, we're therefore just changing the function signaure.
The alternative approach would be to restructure the polkadot
package structure to remove the circular dependency.